feat: proxy deployment to goerli & typescript support
