+++ contracts/package.json
@@ -25,7 +25,9 @@
     "hardhat": "^2.9.1",
     "hardhat-abi-exporter": "^2.10.0",
     "hardhat-deploy": "^0.11.18",
-    "husky": "^8.0.1"
+    "husky": "^8.0.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.3"
   },
   "dependencies": {
     "@openzeppelin/contracts": "^4.7.3",
